refactor(rafty): Refactor state_loop for better node state management
I previously refactored state_loop with a single loop managing different node state.
The problem with that is I noticed that we a node is becoming a leader, release funcs from previous state were only executed when the node was stepping down.
Even worst than that, when a node was starting, the leader was getting stucked(no more append entries send) or was loosing the leadership.
Shutting down the leader was also a mess.
So I combined having a loop for each node states but with the init(), release() and onTimeout() funcs.
Now everything works as it's supposed to be.