refactor(logs): Refactoring all funcs relating to logs management
logsLoop and related funcs implementation was a good idea but in the end, it was extremely slow.
Access to logs could take up to 1.5s with select statement (used to manage graceful shutdown/timeout to release chans and avoid chan leaks).
The worst thing was when adding timeout, we could have many of them and increasing it was obviously not a good idea.
The new implementation is now more cleaner and extremely FASTER. We went from 1.5s AND timeout to 50ns-650µs without ANY timeout.
No data race or race condition has been found with this new implementation.