feat(rafty): Separate logStore and clusterStore
Related to github issue #49

Splitting the store is necessary to have a separation of duty.
Users can still use the same store for both of them but
will use different namespaces/buckets to avoid collisions.
ClusterStore won't be modified that much so caching it won't be that necessary
but still possible.